requirements.txt
pyaudio
google-cloud-speech
openai
